extracted pagination tests from vertical results
